pager item을 설정하는 부분 strategy로 이동
