+++ app/src/main/java/net/jspiner/epub_viewer/ui/reader/strategy/ViewerTypeStrategy.kt
@@ -1,6 +1,7 @@
 package net.jspiner.epub_viewer.ui.reader.strategy
 
 import net.jspiner.epub_viewer.dto.PageInfo
+import net.jspiner.epub_viewer.ui.reader.viewer.EpubPagerAdapter
 import net.jspiner.epub_viewer.ui.reader.viewer.VerticalViewPager
 
 interface ViewerTypeStrategy {
@@ -8,5 +9,12 @@ interface ViewerTypeStrategy {
     fun changeViewPagerOrientation(verticalViewPager: VerticalViewPager)
 
     fun getAllPageCount(pageInfo: PageInfo): Int
+
+    fun setCurrentPagerItem(
+        pager: VerticalViewPager,
+        adapter: EpubPagerAdapter,
+        pageInfo: PageInfo,
+        currentPage: Int
+    )
 
 }
